Sync bitbucket and GitHub

+++ plugins/modules/na_ontap_quota_policy.py
@@ -3,6 +3,10 @@
 # (c) 2019, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
+'''
+na_ontap_quota_policy
+'''
+
 from __future__ import absolute_import, division, print_function
 
 __metaclass__ = type
@@ -13,13 +17,13 @@ ANSIBLE_METADATA = {'metadata_version': '1.1',
 
 DOCUMENTATION = """
 module: na_ontap_quota_policy
-short_description: NetApp Ontap create, rename or delete quota policy
+short_description: NetApp Ontap create, assign, rename or delete quota policy
 extends_documentation_fragment:
     - netapp.ontap.netapp.na_ontap
 version_added: '19.11.0'
 author: NetApp Ansible Team (@carchi8py) <ng-ansibleteam@netapp.com>
 description:
-- Create, rename or delete the quota policy
+- Create, assign, rename or delete the quota policy
 options:
   state:
     description:
@@ -44,6 +48,15 @@ options:
     description:
     - Name of the existing quota policy to be renamed to name.
     type: str
+
+  auto_assign:
+    description:
+      - when true, assign the policy to the vserver, whether it is newly created, renamed, or already exists.
+      - when true, the policy identified by name replaces the already assigned policy.
+      - when false, the policy is created if it does not already exist but is not assigned.
+    type: bool
+    default: true
+    version_added: 20.12.0
 """
 
 EXAMPLES = """
@@ -82,17 +95,18 @@ RETURN = """
 
 import traceback
 
-from ansible_collections.netapp.ontap.plugins.module_utils.netapp_module import NetAppModule
-import ansible_collections.netapp.ontap.plugins.module_utils.netapp as netapp_utils
 from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils._text import to_native
+from ansible_collections.netapp.ontap.plugins.module_utils.netapp_module import NetAppModule
+import ansible_collections.netapp.ontap.plugins.module_utils.netapp as netapp_utils
+import ansible_collections.netapp.ontap.plugins.module_utils.zapis_svm as zapis_svm
 
 HAS_NETAPP_LIB = netapp_utils.has_netapp_lib()
 
 
 class NetAppOntapQuotaPolicy(object):
     """
-        Create, rename or delete a quota policy
+        Create, assign, rename or delete a quota policy
     """
 
     def __init__(self):
@@ -106,7 +120,7 @@ class NetAppOntapQuotaPolicy(object):
             vserver=dict(required=True, type='str'),
             name=dict(required=True, type='str'),
             from_name=dict(required=False, type='str'),
-
+            auto_assign=dict(required=False, type='bool', default=True),
         ))
 
         self.module = AnsibleModule(
@@ -191,25 +205,41 @@ class NetAppOntapQuotaPolicy(object):
                                   exception=traceback.format_exc())
 
     def apply(self):
-        netapp_utils.ems_log_event("na_ontap_efficiency_policy", self.server)
+        netapp_utils.ems_log_event("na_ontap_quota_policy", self.server)
         current = self.get_quota_policy()
         # rename and create are mutually exclusive
         rename, cd_action = None, None
-        if self.parameters.get('from_name'):
+        cd_action = self.na_helper.get_cd_action(current, self.parameters)
+        if cd_action == 'create' and self.parameters.get('from_name'):
+            # create policy by renaming it
             rename = self.na_helper.is_rename_action(self.get_quota_policy(self.parameters['from_name']), current)
-        else:
-            cd_action = self.na_helper.get_cd_action(current, self.parameters)
+            if rename is None:
+                self.module.fail_json(msg='Error renaming quota policy: %s does not exist.' % self.parameters['from_name'])
 
-        if self.na_helper.changed:
-            if self.module.check_mode:
-                pass
-            else:
-                if cd_action == 'create':
-                    self.create_quota_policy()
-                elif cd_action == 'delete':
-                    self.delete_quota_policy()
-                elif rename:
-                    self.rename_quota_policy()
+        # check if policy should be assigned
+        assign_policy = cd_action == 'create' and self.parameters['auto_assign']
+        if cd_action is None and current and self.parameters['auto_assign']:
+            # find out if the existing policy needs to be changed
+            svm = zapis_svm.get_vserver(self.server, self.parameters['vserver'])
+            if svm.get('quota_policy') != self.parameters['name']:
+                assign_policy = True
+                self.na_helper.changed = True
+        if cd_action == 'delete':
+            # can't delete if already assigned
+            svm = zapis_svm.get_vserver(self.server, self.parameters['vserver'])
+            if svm.get('quota_policy') == self.parameters['name']:
+                self.module.fail_json(msg='Error policy %s cannot be deleted as it is assigned to the vserver %s' %
+                                      (self.parameters['name'], self.parameters['vserver']))
+
+        if self.na_helper.changed and not self.module.check_mode:
+            if rename:
+                self.rename_quota_policy()
+            elif cd_action == 'create':
+                self.create_quota_policy()
+            elif cd_action == 'delete':
+                self.delete_quota_policy()
+            if assign_policy:
+                zapis_svm.modify_vserver(self.server, self.module, self.parameters['vserver'], modify=dict(quota_policy=self.parameters['name']))
         self.module.exit_json(changed=self.na_helper.changed)
